Setup instance store through service management
Setup instance store (ephemeral drive) using OS Service Management (init or systemd).
This allows to setup again the instance store after an instance restart.

The creation of the fstab entry is removed: this because the ephemeral drive is replaced by EC2 on restart and the new drive needs to be formatted and mounted.

Resolve https://github.com/aws/aws-parallelcluster/issues/2080

Signed-off-by: Luca Carrogu <carrogu@amazon.com>